# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Checkout Microservice",
        "description": "Checkout Microservice API Documentation",
        "contact": {
            "email": "administrativo@mobiup.com.br"
        },
        "license": {
            "name": "MIT License",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/carts": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get cart",
                "operationId": "224ab5990ba547df8b5718c3dbbeb75c",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create cart",
                "operationId": "816e37674cc93be4dd0836e5e7d95d1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "0060033994"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 10254
                                                },
                                                "weight": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "item1"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart created successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "cartId": 1
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete cart",
                "operationId": "611e4b24c8368b9d34771919b9f6927b",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cart deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/carts/cart-items/{code}": {
            "put": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Update cart item quantity",
                "operationId": "e6d858c5212795f8859eb8f182450bee",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Item code",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "0060033994"
                                    },
                                    "item": {
                                        "properties": {
                                            "quantity": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item quantity updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart item quantity updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Create cart item",
                "operationId": "b8d2b3912eda9012726ae56e59ba2147",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Item code",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "0060033994"
                                    },
                                    "item": {
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "example": 10255
                                            },
                                            "weight": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "item1"
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart item created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Delete cart item",
                "operationId": "6dfc4ebb32c0ee3df793919f77d80462",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Item code",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "0060033994"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Cart item deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/carts/shipping": {
            "put": {
                "tags": [
                    "Cart Shipping"
                ],
                "summary": "Update cart shipping information",
                "operationId": "b6da9e1311b7bc4c373f6fbd031fc5ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "0060033994"
                                    },
                                    "shipping": {
                                        "properties": {
                                            "price": {
                                                "type": "integer",
                                                "example": 30
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart shipping information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart shipping information updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "cartToken": "22bc82df-5750-4116-af17-e258e1d7030c",
                            "itemsSubtotal": "5.00",
                            "taxRate": null,
                            "tax": null,
                            "total": "5.30",
                            "totalCents": 530,
                            "discount": null,
                            "shipping": "0.30",
                            "cartItems": [
                                {
                                    "cartItemToken": "6ae80be6-ef83-402b-b7b8-6cc441f66f14",
                                    "price": "4.00",
                                    "quantity": 4,
                                    "customerNote": null,
                                    "product": {
                                        "name": "item1",
                                        "image": null,
                                        "price": "1.00"
                                    }
                                },
                                {
                                    "cartItemToken": "afac7afa-17fa-484c-bd2e-848ca9c3147d",
                                    "price": "1.00",
                                    "quantity": 1,
                                    "customerNote": null,
                                    "product": {
                                        "name": "item1",
                                        "image": null,
                                        "price": "1.00"
                                    }
                                }
                            ]
                        }
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success message"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200": {
                "description": "Success response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessResponse"
                        }
                    }
                }
            }
        }
    }
}